Debug magnetic GGA
Create a fake GGA functional that gets a magnetic state of Fe.

On a side modification: reshuffle code in such way that xc_functional.hpp is
included only in few cpp files.
