+++ apps/atoms/atom.cpp
@@ -21,6 +21,7 @@
 #include <sirius.hpp>
 #include "mixer/broyden1_mixer.hpp"
 #include "unit_cell/atomic_conf.hpp"
+#include "potential/xc_functional.hpp"
 
 double const rmin{1e-5};
