+++ src/simulation_context.hpp
@@ -32,7 +32,6 @@
 #include "radial/radial_integrals.hpp"
 #include "utils/utils.hpp"
 #include "density/augmentation_operator.hpp"
-#include "potential/xc_functional.hpp"
 #include "gpu/acc.hpp"
 #include "symmetry/check_gvec.hpp"
 #include "symmetry/rotation.hpp"
