+++ src/unit_cell/atom_type.hpp
@@ -28,7 +28,7 @@
 #include "atomic_data.hpp"
 #include "geometry3d.hpp"
 #include "radial/radial_solver.hpp"
-#include "potential/xc_functional_base.hpp"
+//#include "potential/xc_functional_base.hpp"
 #include "simulation_parameters.hpp"
 #include "radial_functions_index.hpp"
 #include "basis_functions_index.hpp"
